// tellie/week7/mission1/src/types/lp.ts
import type { CursorBasedResponse } from "./common";

export type Tag = {
    id: number;
    name: string;
};

export type Likes = {
    id: number;
    userId: number;
    lpId: number;
}

export type Lp = {
    id: number;
    title: string;
    content: string;
    thumbnail: string;
    published: boolean;
    authorId: number;
    author?: {
        id: number;
        name: string;
        profileImage?: string;
    };
    createdAt: Date;
    updatedAt: Date;
    tags: Tag[];
    likes: Likes[];
}

export type Comment = {
    id: number;
    content: string;
    authorId: number;
    lpId: number;
    author: {
        id: number;
        name: string;
        profileImage?: string;
    };
    createdAt: Date;
    updatedAt: Date;
}

export type ResponseLpListDto = CursorBasedResponse<Lp[]>;
export type ResponseLpCommentsDto = CursorBasedResponse<Comment[]>;